Debugged and checked. Judge Full Data behaving poorly.
